server/clean_requirements.txt
agentql==1.11.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
browser-use==0.1.41
cachetools==5.5.2
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
diskcache==5.6.3
docopt==0.6.2
faiss-cpu==1.10.0
fake-http-header==0.3.5
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.59.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
google-genai==1.19.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
graphiti-core==0.11.6
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.40.0
jsonpointer==2.1
langchain==0.3.22
langchain-agentql==1.0.1
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-community==0.3.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-mistralai==0.2.11
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-qwq==0.2.0
langchain-tests==0.3.19
langchain-text-splitters==0.3.7
langsmith==0.3.45
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.3.0
mcp-server-git==2025.1.14
md2pdf==1.0.1
mdurl==0.1.2
mem0ai==0.1.88
mistune==3.1.3
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.28.1
networkx==3.4.2
numpy==2.2.3
ollama==0.4.7
openai==1.75.0
openai-websearch-mcp==0.4.0
orjson==3.10.15
pillow==11.2.1
playwright==1.51.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.21.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydyf==0.11.0
pyee==12.1.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyperclip==1.9.0
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
rubicon-objc==0.5.0
s3transfer==0.11.5
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
secure==1.0.1
sentence-transformers==4.1.0
setuptools==78.1.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.1
syrupy==4.9.1
tavily-python==0.7.3
tenacity==9.0.0
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.6.0
transformers==4.51.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.1
uvicorn==0.34.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.44.0
yarl==1.18.3
zopfli==0.2.3.post1
langgraph-checkpoint-postgres==2.0.0
psycopg[pool]
